+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## TBD
+
+Features:
+- Add "--execute" option to changeset create subcommand.
+
 ## Version 0.3.1 (2017-01-24)
 
 Issues: